# mate-platform-backend/contracts/openapi/generated/bundled.yaml
openapi: 3.1.0
info:
  title: Mate Platform API
  version: 1.0.0
  description: Canonical contract-first API for all Mate Platform domains.
  contact:
    name: Mate Platform API Governance
    email: api-governance@metaplatform.local
  license:
    name: Proprietary
    identifier: LicenseRef-Proprietary
  x-mate-owner: architecture-api-governance
servers:
  - url: /
    description: Gateway-relative production endpoint
security:
  - bearerAuth: []
...
paths:
...
  /api/v1/admin/permissions/assign:
    post:
      summary: Assign permissions to a user or a role
      tags:
        - iam
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssignPayload"
      responses:
        "200":
          description: OK
        "400":
          description: bad type/role_ids combination
      operationId: iamPostAdminPermissionsAssign
      description: Assign permissions to a user or a role
      x-mate-owner: security-iam
      x-mate-permission: iam.post
      x-mate-requirements:
        - FR-IAM-IAMPOSTADMINPERMISSIONSASSIGN
      x-mate-implementation-status: implemented
  /api/v1/admin/permissions/catalog:
    get:
      summary: Permission catalog (resource + actions)
      tags:
        - iam
      responses:
        "200":
          description: OK
        "400":
          $ref: "#/components/responses/Error400"
      operationId: iamGetAdminPermissionsCatalog
      description: Permission catalog (resource + actions)
      x-mate-owner: security-iam
      x-mate-permission: iam.get
      x-mate-requirements:
        - FR-IAM-IAMGETADMINPERMISSIONSCATALOG
      x-mate-implementation-status: implemented
  /api/v1/admin/permissions/matrix:
    get:
      summary: Role x permission matrix
      tags:
        - iam
      parameters:
        - in: query
          name: role_id
          schema:
            type: integer
      responses:
        "200":
          description: OK
        "400":
          $ref: "#/components/responses/Error400"
      operationId: iamGetAdminPermissionsMatrix
      description: Role x permission matrix
      x-mate-owner: security-iam
      x-mate-permission: iam.get
      x-mate-requirements:
        - FR-IAM-IAMGETADMINPERMISSIONSMATRIX
      x-mate-implementation-status: implemented
  /api/v1/admin/permissions/roles:
    get:
      summary: List roles
      tags:
        - iam
      parameters:
        - in: query
          name: keyword
          schema:
            type: string
        - in: query
          name: page
          schema:
            type: integer
            minimum: 1
            default: 1
        - in: query
          name: pageSize
          schema:
            type: integer
            minimum: 1
            maximum: 200
            default: 20
      responses:
        "200":
          description: OK
        "400":
          $ref: "#/components/responses/Error400"
      operationId: iamGetAdminPermissionsRoles
      description: List roles
      x-mate-owner: security-iam
      x-mate-permission: iam.get
      x-mate-requirements:
        - FR-IAM-IAMGETADMINPERMISSIONSROLES
      x-mate-implementation-status: implemented
    post:
      summary: Create role
      tags:
        - iam
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RoleCreate"
      responses:
        "201":
          description: Created
        "409":
          description: Role code already exists
      operationId: iamPostAdminPermissionsRoles
      description: Create role
      x-mate-owner: security-iam
      x-mate-permission: iam.post
      x-mate-requirements:
        - FR-IAM-IAMPOSTADMINPERMISSIONSROLES
      x-mate-implementation-status: implemented
  /api/v1/admin/permissions/roles/{role_id}:
    parameters:
      - in: path
        name: role_id
        required: true
        schema:
          type: integer
    get:
      summary: Role + permission detail
      tags:
        - iam
      responses:
        "200":
          description: OK
        "404":
          description: Not found
      operationId: iamGetAdminPermissionsRolesRoleId
      description: Role + permission detail
      x-mate-owner: security-iam
      x-mate-permission: iam.get
      x-mate-requirements:
        - FR-IAM-IAMGETADMINPERMISSIONSROLESROLEID
      x-mate-implementation-status: implemented
      parameters:
        - name: role_id
          in: path
          required: true
          schema:
            type: string
    put:
      summary: Update role
      tags:
        - iam
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RoleUpdate"
      responses:
        "200":
          description: OK
        "404":
          description: Not found
      operationId: iamPutAdminPermissionsRolesRoleId
      description: Update role
      x-mate-owner: security-iam
      x-mate-permission: iam.put
      x-mate-requirements:
        - FR-IAM-IAMPUTADMINPERMISSIONSROLESROLEID
      x-mate-implementation-status: implemented
      parameters:
        - name: role_id
          in: path
          required: true
          schema:
            type: string
    delete:
      summary: Delete role (rejected if users still assigned)
      tags:
        - iam
      responses:
        "200":
          description: OK
        "404":
          description: Not found
        "409":
          description: Role still has assignments
      operationId: iamDeleteAdminPermissionsRolesRoleId
      description: Delete role (rejected if users still assigned)
      x-mate-owner: security-iam
      x-mate-permission: iam.delete
      x-mate-requirements:
        - FR-IAM-IAMDELETEADMINPERMISSIONSROLESROLEID
      x-mate-implementation-status: implemented
      parameters:
        - name: role_id
          in: path
          required: true
          schema:
            type: string
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
...
    schemas-ErrorResponse:
      type: object
      additionalProperties: false
      required:
        - code
        - message
        - requestId
      properties:
        code:
          type: string
          pattern: ^E[0-9]{3}_[A-Z0-9_]+$
        message:
          type: string
          minLength: 1
        requestId:
          type: string
          minLength: 1
        details:
          type: object
          additionalProperties: true
...
    AssignPayload:
      type: object
      required:
        - type
        - target_id
      properties:
        type:
          type: string
          enum:
            - user
            - role
        target_id:
          type: integer
        permission_ids:
          type: array
          items:
            type: integer
        role_ids:
          type:
            - array
            - "null"
          description: required when type=user
          items:
            type: integer
    RoleCreate:
      type: object
      required:
        - code
        - name
      properties:
        code:
          type: string
          minLength: 2
          maxLength: 64
        name:
          type: string
          minLength: 1
          maxLength: 128
        description:
          type:
            - string
            - "null"
        data_scope:
          type: string
          enum:
            - ALL
            - DEPT
            - DEPT_AND_SUB
            - SELF
            - CUSTOM
          default: SELF
        permission_ids:
          type: array
          items:
            type: integer
    RoleUpdate:
      type: object
      properties:
        name:
          type:
            - string
            - "null"
        description:
          type:
            - string
            - "null"
          maxLength: 8192
        data_scope:
          type:
            - string
            - "null"
          enum:
            - ALL
            - DEPT
            - DEPT_AND_SUB
            - SELF
            - CUSTOM
        permission_ids:
          type:
            - array
            - "null"
          items:
            type: integer
...
  responses:
...
    Error400:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/schemas-ErrorResponse"